day7(반복문)

+++ day07/src/loop/Test07.java
@@ -7,9 +7,9 @@ public class Test07 {
 	public static void main(String[] args) {
 		
 	for(int i=1; i<100; i++) {
-		if((int)(i/10)==3 || (i%10)==3
-		 ||(int)(i/10)==6 || (i%10)==6
-		 ||(int)(i/10)==9 || (i%10)==9) {
+		if((i/10)==3 || (i%10)==3
+		 ||(i/10)==6 || (i%10)==6
+		 ||(i/10)==9 || (i%10)==9) {
 			System.out.println(i);
 			}
 		}
